# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationOut": {
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "invited_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By"
          },
          "last_sent_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sent At"
          },
          "project_ids": {
            "items": {
              "type": "string"
            },
            "title": "Project Ids",
            "type": "array"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "project_ids",
          "created_at",
          "expires_at",
          "last_sent_at",
          "invited_by",
          "accepted_at"
        ],
        "title": "InvitationOut",
        "type": "object"
      },
      "InvitationsOut": {
        "properties": {
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/InvitationOut"
            },
            "title": "Invitations",
            "type": "array"
          },
          "mail_enabled": {
            "title": "Mail Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "mail_enabled",
          "invitations"
        ],
        "title": "InvitationsOut",
        "type": "object"
      },
      "InviteIn": {
        "properties": {
          "deliver": {
            "default": true,
            "title": "Deliver",
            "type": "boolean"
          },
          "emails": {
            "items": {
              "format": "email",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Emails",
            "type": "array"
          },
          "project_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Project Ids",
            "type": "array"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "InviteIn",
        "type": "object"
      },
      "IssuedOut": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mail_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mail Error"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sent": {
            "title": "Sent",
            "type": "boolean"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "expires_at",
          "url",
          "sent"
        ],
        "title": "IssuedOut",
        "type": "object"
      },
...
      "JoinedOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "role"
        ],
        "title": "JoinedOut",
        "type": "object"
      },
...
      "PreviewOut": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "inviter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inviter Name"
          },
          "org_name": {
            "title": "Org Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "org_name",
          "role",
          "email",
          "inviter_name",
          "expires_at"
        ],
        "title": "PreviewOut",
        "type": "object"
      },
...
      "ReissueIn": {
        "properties": {
          "deliver": {
            "default": false,
            "title": "Deliver",
            "type": "boolean"
          }
        },
        "title": "ReissueIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Planora",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/invitations/{token}": {
      "get": {
        "description": "What invitation is in hand — before signing in and before registering.\n\nIt answers anonymous callers too: a person with an unaccepted invitation is\nby definition not in the organization yet, and demanding that they sign in\nbefore learning where they are being invited is asking them to sign\nunseen.",
        "operationId": "preview_invitation_api_invitations__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations/{token}/accept": {
      "post": {
        "description": "Accepts an invitation on behalf of the signed-in person.\n\nA membership appears only here — by a person's explicit action, not by an\naddress matching at registration. The session switches to the new\norganization right away: the person pressed \"accept\" and must end up inside\nrather than looking for it in the switcher.",
        "operationId": "accept_invitation_api_invitations__token__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/api/org/invitations": {
      "get": {
        "description": "The organization's invitations, newest first.\n\nAll of them are returned, accepted ones included: an invitation lives in the\ndatabase after acceptance too — it is the record of who brought whom in. None\nof them carries a plain link in the answer, and none can: the server does not\nremember them.",
        "operationId": "list_invitations_api_org_invitations_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invitations",
        "tags": [
          "invitations"
        ]
      },
      "post": {
        "operationId": "create_invitations_api_org_invitations_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssuedOut"
                  },
                  "title": "Response Create Invitations Api Org Invitations Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invitations",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/org/invitations/{invitation_id}": {
      "delete": {
        "operationId": "revoke_invitation_api_org_invitations__invitation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/org/invitations/{invitation_id}/reissue": {
      "post": {
        "description": "Issues a new link in place of the previous one — also known as \"send again\".\n\nOne action, not two: the previous token dies in both cases, because otherwise\nrevoking an already sent message becomes impossible. Only the delivery\ndiffers — whether a message goes out or the link is copied by hand.",
        "operationId": "reissue_invitation_api_org_invitations__invitation_id__reissue_post",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReissueIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reissue Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
  }
}